requirements.txt
bcrypt==3.2.0
Flask==1.1.2
Flask-Cors==3.0.10
mariadb==1.0.6
PyJWT==1.7.1
gunicorn